requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
